+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Removed double publication enum (P/p)
 - Updated publication directory schema 20230616
 - Changes for CEDAR docs
+- Tweaks for CEDAR release
 
 ## v0.0.15 - 2023-04-04
 - Versioned directory structure schema